refactor(Button): move props type declaration to ui-file
