--- src/shared/ui/button/models/button.ts
@@ -1,9 +1,0 @@
-import { MouseEventHandler, ReactNode } from 'react';
-
-export type ButtonProps = {
-	onClick?: MouseEventHandler<HTMLButtonElement>;
-	children?: ReactNode;
-	disabled?: boolean;
-	variant?: 'default' | 'primary' | 'error' | 'success' | 'info' | 'warning';
-	size?: 'md' | 'sm' | 'lg' | 'icon';
-};
