+++ src/shared/ui/button/ui/button.tsx
@@ -1,7 +1,15 @@
 'use client';
 
-import { ButtonProps } from '../models/button';
+import { MouseEventHandler, ReactNode } from 'react';
 import styles from './button.module.scss';
+
+export type ButtonProps = {
+	onClick?: MouseEventHandler<HTMLButtonElement>;
+	children?: ReactNode;
+	disabled?: boolean;
+	variant?: 'default' | 'primary' | 'error' | 'success' | 'info' | 'warning';
+	size?: 'md' | 'sm' | 'lg' | 'icon';
+};
 
 export function Button(props: ButtonProps) {
 	return (
